Basic navigation working
"describe location" and "move [north|south|east|west]" supported.
Move doesn't update the db yet. It would be better to store the players
separately from the NPCs, as moves will be easier and it will bemuch
easier to find a player's location. This will require a separate
backbone collection for the players.